pyproject.toml
[project]
name = "thesis-kb"
version = "0.1.0"
description = "MCP server for thesis knowledge base with ChromaDB vector search and SQLite FTS5"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "requests",
    "pymupdf",
    "tiktoken",
    "openai",
    "chromadb",
    "numpy",
    "tqdm",
    "flashrank",
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "mcp[cli]>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
